Move set profile info to create session
