+++ client/src/routes/Sessions/components/ProfileInfo.tsx
@@ -7,8 +7,8 @@ import {Spacer16} from '../../../common/components/Spacers/Spacer';
 import Input from '../../../common/components/Typography/TextInput/TextInput';
 import {SPACINGS} from '../../../common/constants/spacings';
 import {userAtom} from '../../../lib/user/state/state';
-import useChangeProfileInfo from '../hooks/useChangeProfileInfo';
-import ProfilePicture from './ProfilePicture';
+import useChangeProfileInfo from '../../Profile/hooks/useChangeProfileInfo';
+import ProfilePicture from '../../../common/components/User/ProfilePicture';
 
 const Container = styled.View({
   alignItems: 'center',
@@ -22,7 +22,7 @@ const ProfilePicutreWrapper = styled.View({
 });
 
 const ProfileInfo = () => {
-  const {t} = useTranslation('Screen.Profile');
+  const {t} = useTranslation('Component.ProfileInfo');
   const user = useRecoilValue(userAtom);
   const {changeProfilePicture, changeProfileName} = useChangeProfileInfo();
   const [displayName, setDisplayName] = useState(user?.displayName ?? '');
@@ -43,11 +43,11 @@ const ProfileInfo = () => {
         autoCapitalize="none"
         autoCorrect={false}
         onSubmitEditing={setProfileName}
-        placeholder={t('userProfile.displayName')}
+        placeholder={t('displayName')}
         onChangeText={setDisplayName}
       />
       <Spacer16 />
-      <Button onPress={setProfileName}>{t('userProfile.saveButton')}</Button>
+      <Button onPress={setProfileName}>{t('cta')}</Button>
     </Container>
   );
 };
