+++ content/src/ui/Component.ProfileInfo.json
@@ -1,0 +1,11 @@
+{
+  "en": {
+    "displayName": "What should we call you?",
+    "cta": "Save"
+  },
+  "sv": {
+    "displayName": "Vad ska vi kalla dig?",
+    "cta": "Spara"
+  },
+  "pt": {}
+}
